# refactor(r): Use `basic_array_stream()` to improve array stream to data.frame conversion