--- a/r/R/convert-array-stream.R
+++ b/r/R/convert-array-stream.R
@@ -82,17 +82,11 @@ convert_array_stream <- function(array_stream, to = NULL, size = NULL, n = Inf) 
   }
 
   # Otherwise, compute the final size, create another array stream,
-  # and call convert_array_stream() with a known size.
-  lengths <- vapply(
-    batches,
-    # Use custom accessor because array$length in a loop is slow
-    function(array) {
-      .Call(nanoarrow_c_array_proxy, array, NULL, FALSE)$length
-    },
-    double(1)
-  )
+  # and call convert_array_stream() with a known size. Using .Call()
+  # directly because we have already type checked the inputs.
+  total_length <- .Call(nanoarrow_c_array_list_total_length, batches)
   basic_stream <- .Call(nanoarrow_c_basic_array_stream, batches, schema, FALSE)
-  convert_array_stream(basic_stream, to = to, size = sum(lengths))
+  convert_array_stream(basic_stream, to = to, size = total_length)
 }
 
 #' @rdname convert_array_stream