--- a/r/tools/make-callentries.R
+++ b/r/tools/make-callentries.R
@@ -74,3 +74,5 @@ stopifnot(str_detect(init, pattern))
 init %>%
   str_replace(pattern, header) %>%
   write_file("src/init.c")
+
+system("clang-format -i src/init.c")